Move AggregateBase into its own file
